requirements.txt
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
pre-commit==3.7.1
psycopg2-binary==2.9.10
python-json-logger==2.0.7
requests==2.32.4
